Moves prefix and suffix to input folder
