+++ components/src/FieldLabel/index.js
@@ -2,5 +2,3 @@ export { default as HelpText } from "./HelpText";
 export { default as RequirementText } from "./RequirementText";
 export { default as FieldLabel } from "./FieldLabel";
 export { default as MaybeFieldLabel } from "./MaybeFieldLabel";
-export { default as Suffix } from "./Suffix";
-export { default as Prefix } from "./Prefix";
