+++ components/src/Input/Input.js
@@ -5,7 +5,8 @@ import { transparentize } from "polished";
 import { space } from "styled-system";
 import { Field } from "../Form";
 import { Flex } from "../Flex";
-import { MaybeFieldLabel, Prefix, Suffix } from "../FieldLabel";
+import { MaybeFieldLabel } from "../FieldLabel";
+import { Prefix, Suffix } from "../Input";
 import { InlineValidation } from "../Validation";
 import theme from "../theme";
 import { subPx } from "../utils";
