+++ components/src/Input/index.js
@@ -1,1 +1,3 @@
 export { default as Input } from "./Input";
+export { default as Suffix } from "../Input/Suffix";
+export { default as Prefix } from "../Input/Prefix";
