+++ components/src/index.js
@@ -13,8 +13,8 @@ export { InlineValidation, HeaderValidation } from "./Validation";
 export { Alert } from "./Alert";
 export { Branding, BrandingText } from "./Branding";
 export { DropdownMenu, DropdownLink, DropdownButton, DropdownItem } from "./DropdownMenu";
-export { HelpText, RequirementText, FieldLabel, MaybeFieldLabel, Suffix, Prefix } from "./FieldLabel";
-export { Input } from "./Input";
+export { HelpText, RequirementText, FieldLabel, MaybeFieldLabel } from "./FieldLabel";
+export { Input, Suffix, Prefix } from "./Input";
 export { NavBarSearch } from "./NavBarSearch";
 export { NavBar } from "./NavBar";
 export { Checkbox, CheckboxGroup } from "./Checkbox";
